Replace "-gnatk8" switch with explicit Naming package mapping
GNAT supports the notion of so-called filename krunching [sic] for Ada
sources, but there appears to be no simple way to enforce this upon the
closure of sources within a given GPRBuild project. This poses an
issue for library clients that do not use krunching themselves.

One solution would be to expand the filenames, but that would simply
transfer the burden to users of, e.g., GNAT on DJGPP. As an
alternative, it is possible to explicitly specify a mapping with
GPRBuild's package Naming between Ada specification and body names and
filenames.

Thanks to @mgrojo for revealing this issue.